requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
mediapipe==0.10.14
opencv-python-headless==4.8.0.76
tensorflow-cpu==2.13.0
joblib==1.3.1
python-multipart==0.0.6
scikit-learn==1.3.0
